Add docs input `RadioList::class`.
